[refactor/#5]: dto -> request & response 패키지 분리 및 리팩토링
