+++ src/main/java/fisa/woorizip/backend/house/controller/HouseController.java
@@ -1,7 +1,7 @@
 package fisa.woorizip.backend.house.controller;
 
-import fisa.woorizip.backend.house.dto.MapFilterRequest;
-import fisa.woorizip.backend.house.dto.ShowMapResponse;
+import fisa.woorizip.backend.house.dto.request.MapFilterRequest;
+import fisa.woorizip.backend.house.dto.response.ShowMapResponse;
 import fisa.woorizip.backend.house.service.HouseService;
 
 import jakarta.validation.Valid;
@@ -9,7 +9,6 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
