+++ src/main/java/fisa/woorizip/backend/house/dto/response/HouseContentResponse.java
@@ -1,4 +1,4 @@
-package fisa.woorizip.backend.house.dto;
+package fisa.woorizip.backend.house.dto.response;
 
 import fisa.woorizip.backend.house.domain.House;
 import fisa.woorizip.backend.house.domain.HouseType;
