+++ src/main/java/fisa/woorizip/backend/house/dto/response/ShowMapResponse.java
@@ -1,6 +1,7 @@
-package fisa.woorizip.backend.house.dto;
+package fisa.woorizip.backend.house.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import fisa.woorizip.backend.house.dto.HouseAddressType;
 import lombok.Getter;
 import org.springframework.data.domain.Slice;
