+++ src/main/java/fisa/woorizip/backend/house/service/HouseService.java
@@ -1,13 +1,8 @@
 package fisa.woorizip.backend.house.service;
 
-import fisa.woorizip.backend.facility.domain.Category;
-import fisa.woorizip.backend.house.domain.HouseType;
-import fisa.woorizip.backend.house.domain.HousingExpenses;
-import fisa.woorizip.backend.house.dto.MapFilterRequest;
-import fisa.woorizip.backend.house.dto.MapLevel;
-import fisa.woorizip.backend.house.dto.ShowMapResponse;
+import fisa.woorizip.backend.house.dto.request.MapFilterRequest;
+import fisa.woorizip.backend.house.dto.response.ShowMapResponse;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface HouseService {
