+++ src/main/java/fisa/woorizip/backend/house/service/HouseServiceImpl.java
@@ -1,10 +1,7 @@
 package fisa.woorizip.backend.house.service;
 
-import fisa.woorizip.backend.facility.domain.Category;
-import fisa.woorizip.backend.house.domain.HouseType;
-import fisa.woorizip.backend.house.domain.HousingExpenses;
-import fisa.woorizip.backend.house.dto.MapFilterRequest;
-import fisa.woorizip.backend.house.dto.ShowMapResponse;
+import fisa.woorizip.backend.house.dto.request.MapFilterRequest;
+import fisa.woorizip.backend.house.dto.response.ShowMapResponse;
 import fisa.woorizip.backend.house.repository.HouseRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -12,7 +9,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Service
